# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sandbox administration and multi-tenant provisioning API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants/{tenantId}/website-preview:
    get:
      operationId: getTenantWebsitePreview
      tags: [Tenants]
      summary: Get an authenticated operator-only read-only website preview without activation
      parameters:
        - name: tenantId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Authenticated launcher; issues a 15-minute tenant-scoped cookie and requires fresh Super Admin authorization on every preview request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebsitePreview"
...
  /tenants/{tenantId}/administrators:
    parameters: [{$ref: "#/components/parameters/TenantId"}]
    get:
      operationId: listTenantAdministrators
      tags: [tenants]
      responses:
        "200":
          description: Assigned Client Admins
          content:
            application/json:
              schema: {type: array, items: {$ref: "#/components/schemas/ResourceItem"}}
    post:
      operationId: assignTenantAdministrator
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TenantAdministratorInput"}
      responses:
        "201":
          description: Client Admin assigned by Super Admin
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MutationStatus"}
  /tenants/{tenantId}/administrators/{userId}:
    put:
      operationId: setTenantAdministratorStatus
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - {name: userId, in: path, required: true, schema: {type: string, pattern: "^user_[a-zA-Z0-9]+$"}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/TenantAdministratorStatusInput"}
      responses:
        "200":
          description: Client Admin activation changed by Super Admin
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MutationStatus"}
  /tenants/{tenantId}/exchange/preview-integrations:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: getExchangePreviewIntegrations
      tags: [exchange]
      responses:
        "200":
          description: Optional private sandbox settings, off by default and never executed
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExchangePreviewIntegrations"}
    put:
      operationId: saveExchangePreviewIntegrations
      tags: [exchange]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ExchangePreviewInput"}
      responses:
        "200":
          description: Stored preview settings only; no API access, webhook delivery or RPC connection
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ExchangePreviewIntegrations"}
  /tenants/{tenantId}/products/{moduleKey}/configuration:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - {name: moduleKey, in: path, required: true, schema: {type: string, maxLength: 64, pattern: "^[a-z][a-z0-9_]*$"}}
    get:
      operationId: getProductConfiguration
      tags: [tenants]
      responses:
        "200":
          description: Tenant product settings only, never execution
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProductConfiguration"}
    put:
      operationId: setProductConfiguration
      tags: [tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/ProductConfiguration"}
      responses:
        "200":
          description: Stored configuration
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ProductConfiguration"}
...
  /tenants/{tenantId}/staff/{userId}/permissions:
    put:
      operationId: setStaffPermissions
      tags: [tenants]
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - {name: userId, in: path, required: true, schema: {type: string, pattern: "^user_[a-zA-Z0-9]+$"}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/StaffPermissions"}
      responses:
        "200":
          description: Staff permissions saved
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MutationStatus"}
  /tenants/{tenantId}/domain-verification:
    get:
      operationId: getDomainVerification
      tags: [tenants]
      parameters: [{$ref: "#/components/parameters/TenantId"}]
      responses:
        "200":
          description: DNS ownership challenge (not a hosting connection)
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DomainVerification"}
    post:
      operationId: verifyTenantDomain
      tags: [tenants]
      parameters: [{$ref: "#/components/parameters/TenantId"}]
      responses:
        "200":
          description: Ownership verified
          content:
            application/json:
              schema: {$ref: "#/components/schemas/DomainVerification"}
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List tenants visible to the current administrator
      responses:
        "200":
          description: Tenant list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TenantSummary"
        "401":
          description: Sign-in required
        "403":
          description: Administrator access required
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a draft tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Draft tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid tenant input
        "401":
          description: Sign-in required
        "403":
          description: Super administrator access required
        "409":
          description: Tenant slug is already in use
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get one tenant and its provisioning state
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
  /tenants/{tenantId}/brand:
    put:
      operationId: updateTenantBrand
      tags: [tenants]
      summary: Set the tenant brand
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BrandInput"
      responses:
        "200":
          description: Brand saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid brand settings
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
  /tenants/{tenantId}/domain:
    put:
      operationId: updateTenantDomain
      tags: [tenants]
      summary: Set or clear a tenant domain
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DomainInput"
      responses:
        "200":
          description: Domain settings saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid domain
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
        "409":
          description: Domain is already assigned
  /tenants/{tenantId}/modules:
    put:
      operationId: updateTenantModules
      tags: [tenants]
      summary: Set enabled product modules for a tenant
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantModulesInput"
      responses:
        "200":
          description: Module entitlements saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid module selection
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
  /tenants/{tenantId}/assets-networks:
    put:
      operationId: updateTenantAssetsNetworks
      tags: [tenants]
      summary: Set supported sandbox assets and networks for a tenant
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetsNetworksInput"
      responses:
        "200":
          description: Supported assets and networks saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid catalog selection
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
  /tenants/{tenantId}/configuration:
    put:
      operationId: updateTenantConfiguration
      tags: [tenants]
      summary: Set sandbox-only tenant configuration
      parameters:
        - $ref: "#/components/parameters/TenantId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantConfigurationInput"
      responses:
        "200":
          description: Configuration saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Invalid configuration
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
  /tenants/{tenantId}/launch:
    post:
      operationId: activateTenant
      tags: [tenants]
      summary: Activate a fully provisioned tenant in sandbox mode
      parameters:
        - $ref: "#/components/parameters/TenantId"
      responses:
        "200":
          description: Tenant activated in sandbox mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          description: Required provisioning steps are incomplete
        "401":
          description: Sign-in required
        "403":
          description: Tenant access denied
        "404":
          description: Tenant not found
...
  /public/sites/{slug}:
    get:
      operationId: getPublicSite
      tags: [catalog]
      parameters: [{name: slug, in: path, required: true, schema: {type: string, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"}}]
      responses:
        "200": {description: Allowlisted public tenant website configuration, content: {application/json: {schema: {$ref: "#/components/schemas/PublicSite"}}}}
...
components:
  parameters:
...
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    WebsitePreview:
      type: object
      required: [url, expiresAt]
      properties:
        url: { type: string }
        expiresAt: { type: string, format: date-time }
...
    TenantAdministratorInput:
      type: object
      properties:
        userId: {type: string, pattern: "^user_[a-zA-Z0-9]+$"}
        label: {type: string, minLength: 2, maxLength: 120}
      required: [userId, label]
    TenantAdministratorStatusInput:
      type: object
      properties: {active: {type: boolean}}
      required: [active]
    ExchangePreviewSettings:
      type: object
      additionalProperties: false
      required: [api, webhooks, rpc]
      properties:
        api:
          type: object
          additionalProperties: false
          required: [enabled, label, baseUrl]
          properties:
            enabled: {type: boolean}
            label: {type: string, maxLength: 120}
            baseUrl: {type: string, maxLength: 2048}
        webhooks:
          type: object
          additionalProperties: false
          required: [enabled, label, endpointUrl, events]
          properties:
            enabled: {type: boolean}
            label: {type: string, maxLength: 120}
            endpointUrl: {type: string, maxLength: 2048}
            events:
              type: array
              maxItems: 3
              uniqueItems: true
              items: {type: string, enum: [order.created, order.status_changed, quote.created]}
        rpc:
          type: object
          additionalProperties: false
          required: [enabled, label, endpointUrl, networkName]
          properties:
            enabled: {type: boolean}
            label: {type: string, maxLength: 120}
            endpointUrl: {type: string, maxLength: 2048}
            networkName: {type: string, maxLength: 120}
    ExchangePreviewInput:
      type: object
      additionalProperties: false
      required: [settings]
      properties:
        settings: {$ref: "#/components/schemas/ExchangePreviewSettings"}
        modules:
          type: array
          minItems: 1
          maxItems: 3
          uniqueItems: true
          items: {type: string, enum: [api, webhooks, rpc]}
    ExchangePreviewIntegrations:
      type: object
      required: [settings, sandboxOnly, executionStatus]
      properties:
        settings: {$ref: "#/components/schemas/ExchangePreviewSettings"}
        sandboxOnly: {type: boolean, enum: [true]}
        executionStatus: {type: string, enum: [configuration_only]}
    ProductConfiguration:
      type: object
      properties:
        configuration: {type: object, additionalProperties: {}}
      required: [configuration]
...
    StaffPermissions:
      type: object
      properties:
        permissions:
          type: array
          uniqueItems: true
          maxItems: 4
          items: {type: string, enum: [branding.manage, domains.manage, configuration.manage, resources.manage]}
      required: [permissions]
    DomainVerification:
      type: object
      properties:
        domain: {type: ["string", "null"]}
        status: {type: string, enum: [unconfigured, unverified, verified]}
        txtName: {type: ["string", "null"]}
        txtValue: {type: ["string", "null"]}
        hostingConnected: {type: boolean}
      required: [domain, status, txtName, txtValue, hostingConnected]
...
    TenantStatus:
      type: string
      enum: [draft, active, suspended]
    ProvisioningStep:
      type: string
      enum: [brand, domain, modules, assets_networks, configuration, ready]
    ModuleKey:
      type: string
      maxLength: 64
      pattern: "^[a-z][a-z0-9_]*$"
...
    AssetNetwork:
      type: object
      properties:
        assetId: { type: string }
        assetNetworkId: {type: string}
        symbol: { type: string }
        name: { type: string }
        networkId: { type: string }
        networkName: { type: string }
        testnet: { type: boolean }
        logoUrl: {type: ["string", "null"]}
        decimals: {type: integer}
      required: [assetId, symbol, name, networkId, networkName, testnet]
...
    TenantInput:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 120 }
        slug: { type: string, minLength: 2, maxLength: 48, pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$" }
        planId: { type: string, format: uuid }
      required: [name, slug, planId]
    BrandInput:
      type: object
      properties:
        brandName: { type: string, minLength: 2, maxLength: 120 }
        logoUrl: { type: ["string", "null"], maxLength: 2048 }
        primaryColor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        accentColor: { type: string, pattern: "^#[0-9A-Fa-f]{6}$" }
        themeMode: { type: string, enum: [light, dark, system] }
        defaultLanguage: { type: string, minLength: 2, maxLength: 12 }
        supportedLanguages:
          type: array
          items: { type: string, minLength: 2, maxLength: 12 }
      required: [brandName, logoUrl, primaryColor, accentColor, themeMode, defaultLanguage, supportedLanguages]
    DomainInput:
      type: object
      properties:
        domain: { type: ["string", "null"], maxLength: 253 }
      required: [domain]
    TenantModulesInput:
      type: object
      properties:
        moduleKeys:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/ModuleKey" }
      required: [moduleKeys]
    AssetsNetworksInput:
      type: object
      properties:
        assetNetworkIds:
          type: array
          uniqueItems: true
          items: { type: string }
      required: [assetNetworkIds]
    TenantConfigurationInput:
      type: object
      properties:
        environment: { type: string, enum: [sandbox] }
        exchangeEnabled: { type: boolean }
        paymentsEnabled: { type: boolean }
        allowGuestCheckout: { type: boolean }
      required: [environment, exchangeEnabled, paymentsEnabled, allowGuestCheckout]
    TenantSummary:
      type: object
      properties:
        exchangeProvisioned: {type: boolean}
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        brandName: { type: string }
        domain: { type: ["string", "null"] }
        status: { $ref: "#/components/schemas/TenantStatus" }
        enabledModules:
          type: array
          items: { $ref: "#/components/schemas/ModuleKey" }
        provisioningStep: { $ref: "#/components/schemas/ProvisioningStep" }
        environment: { type: string, enum: [sandbox] }
        createdAt: { type: string, format: date-time }
      required: [id, name, slug, brandName, domain, status, enabledModules, provisioningStep, environment, createdAt]
    Tenant:
      allOf:
        - $ref: "#/components/schemas/TenantSummary"
        - type: object
          properties:
            logoUrl: { type: ["string", "null"] }
            primaryColor: { type: string }
            accentColor: { type: string }
            themeMode: { type: string, enum: [light, dark, system] }
            defaultLanguage: { type: string }
            supportedLanguages:
              type: array
              items: { type: string }
            assetNetworkIds:
              type: array
              items: { type: string }
            exchangeEnabled: { type: boolean }
            paymentsEnabled: { type: boolean }
            allowGuestCheckout: { type: boolean }
            configurationComplete: { type: boolean }
            activationBlockers:
              type: array
              items: {type: string}
            websiteSettings: {$ref: "#/components/schemas/WebsiteSettings"}
          required: [logoUrl, primaryColor, accentColor, themeMode, defaultLanguage, supportedLanguages, assetNetworkIds, exchangeEnabled, paymentsEnabled, allowGuestCheckout, configurationComplete]
...
    WebsiteSettings:
      type: object
      properties:
        navigation:
          type: array
          maxItems: 12
          items:
            type: object
            properties:
              key: {type: string, enum: [exchange, payments, telegram, how, developers, about, faq]}
              label: {type: string, minLength: 1, maxLength: 60}
              visible: {type: boolean}
            required: [key, label, visible]
        secondaryColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        templateKey: {type: string, enum: [standard-exchange]}
        websiteName: {type: string, minLength: 2, maxLength: 120}
        sandboxLabel: {type: string, enum: [Sandbox, Sandbox Demo]}
        glowColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        surfaceStyle: {type: string, enum: [solid, glass]}
        borderRadius: {type: string, enum: [sharp, soft, rounded]}
        faq: {type: array, maxItems: 16, items: {$ref: "#/components/schemas/WebsiteFaq"}}
        faviconUrl: {type: ["string", "null"], maxLength: 2048}
        fontKey: {type: string, enum: [system, inter, manrope, dm-sans, space-grotesk]}
        heroTitle: {type: string, minLength: 2, maxLength: 180}
        heroSubtitle: {type: string, maxLength: 1000}
        supportEmail: {type: ["string", "null"], maxLength: 254}
        supportUrl: {type: ["string", "null"], maxLength: 2048}
        supportDetails: {type: string, maxLength: 2000}
        socialLinks: {type: array, maxItems: 12, items: {$ref: "#/components/schemas/SiteLink"}}
        footerText: {type: string, maxLength: 2000}
        privacyContent: {type: string, maxLength: 20000}
        termsContent: {type: string, maxLength: 20000}
      required: [secondaryColor, faviconUrl, fontKey, heroTitle, heroSubtitle, supportEmail, supportUrl, supportDetails, socialLinks, footerText, privacyContent, termsContent]
    WebsiteFaq:
      type: object
      properties:
        question: {type: string, minLength: 2, maxLength: 180}
        answer: {type: string, minLength: 2, maxLength: 2000}
      required: [question, answer]
    SiteLink:
      type: object
      properties: {label: {type: string, minLength: 1, maxLength: 80}, url: {type: string, maxLength: 2048}}
      required: [label, url]
...
    ResourceItem:
      type: object
      properties:
        id: {type: string}
        label: {type: string}
        reference: {type: ["string", "null"]}
        status: {type: string}
        permissions: {type: array, items: {type: string}}
      required: [id, label, reference, status]
...
    MutationStatus:
      type: object
      properties: {ok: {type: boolean}}
      required: [ok]
    PublicSite:
      type: object
      properties:
        tenantSlug: {type: string}
        brandName: {type: string}
        logoUrl: {type: ["string", "null"]}
        primaryColor: {type: string}
        accentColor: {type: string}
        themeMode: {type: string, enum: [light, dark, system]}
        domain: {type: ["string", "null"]}
        sandboxOnly: {type: boolean}
        websiteSettings: {$ref: "#/components/schemas/WebsiteSettings"}
        features: {type: object, additionalProperties: {type: boolean}}
        assets: {type: array, items: {$ref: "#/components/schemas/AssetNetwork"}}
      required: [tenantSlug, brandName, logoUrl, primaryColor, accentColor, themeMode, domain, sandboxOnly, websiteSettings, features, assets]
...